[BUGFIX] Select value displayed instead of label
When the field is not editable and therefore rendered in 'show' layout, the select is not rendered and the plain value is displayed instead of the label. This is an issue especially when non-meaningful values are used, like integers.
Fixed by removing the 'not readonly' check, so the select is always rendered, and applying the check to the select configuration instead.