fix: bridge objects not found
Caused by unpublished state which should be updated based on the published
state of the base data object, but isn't updated in some cases.
Since it's not really necessary to keep bridge objects unpublished, publish
them by default and remove the published check to prevent any issues with
existing setups.